backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.0
psycopg2==2.9.6
python-jose==3.3.0
bcrypt==4.0.1
pydantic==2.3.0
python-dotenv==1.0.0